Resolved more file name collisions
